Reorganize
And by that i mean move definitions into classes and create
headers/implementation files. Main game things are in the Game class,
player is in the Player class, both in their own files.
